requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-Migrate==4.0.5
pandas==2.0.3
openpyxl==3.1.2
python-dotenv==1.0.0
psycopg2-binary==2.9.7
boto3==1.28.62
Werkzeug==2.3.7
